chore: updating to match the refactor at 0.23.0
