+++ circuit/Nargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "noirjs_demo"
+name = "circuit"
 type = "bin"
 authors = [""]
-compiler_version = "0.17.0"
+compiler_version = ">=0.22.0"
 
 [dependencies]
